+++ cloc.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
+set -e
+
 DIR="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 #DIR=`dirname $0`